refactor: extract types
