+++ src/components/items/Items.tsx
@@ -1,11 +1,10 @@
 import React from 'react';
-import { IItem } from '../item-details/ItemDetails';
 import styles from './items.module.css';
+import {IItem} from "../../types/types";
 
 interface IProps {
   items: IItem[];
   onItemClick: (url: string) => void;
-  // setItemShown: (boolean) => void;
 }
 
 function Items({ items, onItemClick }: IProps) {
