+++ src/types/types.ts
@@ -1,0 +1,25 @@
+export interface IItem {
+    name: string;
+    height: string;
+    mass: string;
+    hair_color: string;
+    skin_color: string;
+    eye_color: string;
+    birth_year: string;
+    gender: string;
+    homeworld: string;
+    films: string[];
+    species: [];
+    vehicles: string[];
+    starships: string[];
+    created: string;
+    edited: string;
+    url: string;
+}
+
+export interface IResponse {
+    count: number;
+    next: string;
+    previous: string;
+    results: IItem[];
+}
